Move some types from `plumbing.rs` to `job.rs`
Because `ActiveKeyStatus` wraps `QueryJob`, and `QueryState` is a
collection of `ActiveKeyStatus`es.
